1647: Fix Consent UI loading issue
Fixes an issue where the redirect_uri to which the user's browser
should be redirected to after consent was obtained from the
redirect_uri parameter that was a part of the authorize request found in
the consent_request JWT's consentApprovalRedirectUri claim.

The RCS UI should use the redirect_uri from the request jwt that was
provided as a parameter in that consentApprovalRedirectUri claim.

Issue: https://github.com/SecureApiGateway/SecureApiGateway/issues/1647